Move some of the new code to fastforward.go
